requirements.txt
flask
nltk
scikit-learn
gensim
numpy
rank_bm25
PyPDF2
google-generativeai
textwrap